Add CI job that checks if extension is properly updated (#104)
* Recreate all functions in update path

Previously, functions were added in the extension schema, but weren't
added in update paths. This means that if the extension was installed in
a previous version and updated with `ALTER EXTENSION`, it wouldn't
include the new functions. This commit CREATES or REPLACES all
functions exported by pgmq, to guarantee that users updating to pgmq
0.22 will have all functions currently exported.

Add 0.20.0 -> 0.21.0 update path

Fix for signature change in 0.13

* Attempt to make it work

* Use CI user

* Change strategy for branch name

* Fix migrations to make them compatible with no pgmq_meta

This scenario can happen if you installed an old version and never created a queue

* Query fixes

* Fix sql

* Remove TODO comment

* Fix 0.23 migration

* )

* Create queues before updating

* Add client tests

* Alter deleted_at index if exists

* Change strategy for database url

* Fix database url

* Update extension_upgrade.yml

* Update extension_upgrade.yml

* Update extension_upgrade.yml

* Install pg partman

* explicit signature on pgmq-rs send

* drop old send

* test sdk in other workflow

* run pgmq-rs migrate on test db

---------

Co-authored-by: Adam Hendel <ChuckHend@users.noreply.github.com>
Co-authored-by: Adam Hendel <adam@tembo.io>