Fix off by one in metrics (again)
The fix on commit 6d5d695ccba0687839624a41f1878465e6ac7fdf was naive.

When the seq wasn't used, `last_value` is 1, and `is_called` is false.
When the seq was used once, `last_value` is 1, and `is_called` is true.
When the seq was used more than once, `last_value` is N, and `is_called`
is true.