Use sqlx without TLS feature
The pgmq extension doesn't use TLS at all - the resulting so library is
not linked against libcrypto, i.e. the part of the code using sqlx
connect is not used here.

The TLS support adds unnecessary dependencies that increases the
build time and make building on some platform difficult (due to
openssl dependency or lack of rustls/ring support for these platforms).

runtime-tokio is a subset of runtime-tokio-native-tls, so if some
projects that depend on pgmq-core or pgmq-rs need TLS, they can easily
enable runtime-tokio-native-tls on their side.