Add PGXN release workflow
PGXN requires a `META.json` file, so add it and a `make` target to fill
in the version from `Cargo.toml`. Add `.gitattributes` to prevent `git
archive` from archiving some files (should more be omitted from the
release? What about the `.cargo` directory?) and
`.github/workflows/pgxn-release.yml` to do the actual release when a tag
is pushed.

Also tweak the `README` a bit to address some inconsistencies, and to
include a PGXN version badge.