Generate META.json
Rather than modify it in the repository and potentially create weird
conflicts. Generate it from an input file, instead. Use the new
`GIT_BUNDLE_OPTS` variable for the pgxn/pgxn-tools container to add the
generated `META.json` file to the bundle. It would otherwise be excluded
because `git archive` includes only committed files.