fix: Override types of timestamp columns to avoid `time` conflict
Problem
-------
If a consumer has both the `time` and `chrono` features of `sqlx`
enabled, `sqlx` will default to using `OffsetDateTime` (from `time`)
for timestamp fields instead of `DateTime<Utc>` (from `chrono`).
It's easy to accidentally encounter this situation, either by directly
enabling `time`, e.g., if the consumer prefers `time` over `chrono`, or
if the `time` feature is enabled by a transitive dependency. When this
situation occurs, the `pg_ext::PGMQueueMeta` struct can not be built
from the sqlx query because an `OffsetDateTime` was returned instead of
a `DateTime<Utc>`, causing a compile error.

Solution
--------
`sqlx` provides a way to override the type of an output column:
https://docs.rs/sqlx/latest/sqlx/macro.query.html#force-a-differentcustom-type.
Update the failing queries in the `pg_ext` to use this approach.
Unfortunately, I think this requires using the more verbose approach of
naming each column to be returned instead of using `*`.

Testing
-------
- I ran the examples and they appeared to run successfully
- I enabled the `time` feature of `sqlx` in the `dev-dependencies` to
  ensure this case of enabling `chrono` and `time` is supported, which
  can be confirmed by compiling examples/tests.

Other
-----
I also updated the sqlx query caches. I noticed there was an unused
column called `headers` in some of the queries that used a `*` selector.
I can add the column back, but it didn't appear to be used, so I left it
out.